scripts/twitter_slicer/requirements.txt
python-dateutil>=2.8
openai>=1.0
